Refaactored the installation of the /etc/buildkite-agent folder.

+++ packaging/linux/scripts/after-install.sh
@@ -14,17 +14,44 @@ cat <<"TXT"
 
 TXT
 
+# Create the /etc/buildkite-agent folder if it's not there
+if [ ! -d /etc/buildkite-agent ]; then
+  mkdir -p /etc/buildkite-agent
+fi
+
+# Install the config template if it's not there
+if [ ! -f /lib/systemd/system/buildkite-agent.service ]; then
+  cp /usr/share/buildkite-agent/buildkite-agent.cfg /lib/systemd/system/buildkite-agent.service
+fi
+
+# Install the bootstrap.sh file if it doesn't exist
+if [ ! -f /etc/buildkite-agent/bootstrap.sh ]; then
+  cp /usr/share/buildkite-agent/bootstrap.sh /etc/buildkite-agent/bootstrap.sh
+fi
+
+# Copy the hooks if they aren't there
+if [ ! -d /etc/buildkite-agent/hooks ]; then
+  cp -r /usr/share/buildkite-agent/hooks /etc/buildkite-agent
+fi
+
+# Install the relevant system process
 if command -v systemctl > /dev/null; then
-  # Install systemd service
-  cp /usr/share/buildkite-agent/systemd/buildkite-agent.service /lib/systemd/system/buildkite-agent.service
+  if [ ! -f /lib/systemd/system/buildkite-agent.service ]; then
+    cp /usr/share/buildkite-agent/systemd/buildkite-agent.service /lib/systemd/system/buildkite-agent.service
+  fi
+
   START_COMMAND="sudo systemctl enable buildkite-agent && sudo systemctl start buildkite-agent"
 elif [ -d /etc/init ]; then
-  # Install upstart service
-  cp -r /usr/share/buildkite-agent/upstart/buildkite-agent.conf /etc/init/buildkite-agent.conf
+  if [ ! -f /etc/init/buildkite-agent.conf ]; then
+    cp -r /usr/share/buildkite-agent/upstart/buildkite-agent.conf /etc/init/buildkite-agent.conf
+  fi
+
   START_COMMAND="sudo service buildkite-agent start"
 elif [ -d /etc/init.d ]; then
-  # Install old-school init.d script
-  cp -r /usr/share/buildkite-agent/lsb/buildkite-agent.conf /etc/init.d/buildkite-agent
+  if [ ! -f /etc/init.d/buildkite-agent ]; then
+    cp -r /usr/share/buildkite-agent/lsb/buildkite-agent.conf /etc/init.d/buildkite-agent
+  fi
+
   START_COMMAND="sudo /etc/init.d/buildkite-agent start"
 else
   # If all the others fails, warn them and just let them run it the old
